feat: add new design
