+++ package.json
@@ -25,6 +25,7 @@
     "astro": "2.0.6",
     "js-confetti": "0.11.0",
     "react": "18.2.0",
+    "react-countdown": "2.3.5",
     "react-dom": "18.2.0",
     "sweetalert2": "11.7.1",
     "tailwindcss": "3.2.4"
